+++ iOSSettingsScreen HW13/Resources/App/SceneDelegate.swift
@@ -1,0 +1,23 @@
+//
+//  SceneDelegate.swift
+//  iOSSettingsScreen HW13
+//
+//  Created by Артур Горлов on 28.08.2022.
+//
+
+import UIKit
+
+class SceneDelegate: UIResponder, UIWindowSceneDelegate {
+
+    var window: UIWindow?
+
+    func scene(_ scene: UIScene, willConnectTo session: UISceneSession, options connectionOptions: UIScene.ConnectionOptions) {
+        guard let windowScene = (scene as? UIWindowScene) else { return }
+        window = UIWindow(windowScene: windowScene)
+        let settingsViewController = SettingsViewController()
+        let navigationController = UINavigationController(rootViewController: settingsViewController)
+        window?.rootViewController = navigationController
+        window?.makeKeyAndVisible()
+    }
+}
+
