+++ iOSSettingsScreen HW13/Sources/Settings/SettingsViewController.swift
@@ -1,5 +1,5 @@
 //
-//  ViewController.swift
+//  SettingsViewController.swift
 //  iOSSettingsScreen HW13
 //
 //  Created by Артур Горлов on 28.08.2022.
@@ -7,13 +7,10 @@
 
 import UIKit
 
-class ViewController: UIViewController {
+class SettingsViewController: UIViewController {
 
     override func viewDidLoad() {
         super.viewDidLoad()
-        // Do any additional setup after loading the view.
     }
-
-
 }
